requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-heroku==0.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
